+++ mitra_activitypub/src/vocabulary.rs
@@ -29,6 +29,7 @@ pub const AUDIO: &str = "Audio";
 pub const DOCUMENT: &str = "Document";
 pub const IMAGE: &str = "Image";
 pub const NOTE: &str = "Note";
+pub const QUESTION: &str = "Question";
 pub const TOMBSTONE: &str = "Tombstone";
 pub const VIDEO: &str = "Video";