+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Implemented remote interaction with posts.
 - Added portable outbox view.
 - Support changing mitractl log level using `--log-level` parameter.
+- Insert poll results into post as text.
 
 ### Changed